fix color picker
